refactor: change name of list map to layout
