+++ components/cards/ExploreGenreCard.tsx
@@ -1,0 +1,9 @@
+import { View, Text } from "react-native";
+
+export default function ExploreGenreCard() {
+  return (
+    <View>
+      <Text>This is a genreCard</Text>
+    </View>
+  );
+}
